[T1-556] feat: UserTable useQuery로 수정
