+++ packages/wms/src/features/user/model/queryHooks.ts
@@ -1,61 +1,53 @@
-import { Sort } from "@shared/model";
-import { useMutation, useQuery, useSuspenseQuery } from "@tanstack/react-query";
-import {
-  getRoles,
-  getUserCount,
-  getUsers,
-  getUsersPaged,
-  updateUserActive,
-} from "../api/UserApi";
-import { UserFilter } from "@/features";
-
-export const useUserCount = () => {
-  return useSuspenseQuery({
-    queryKey: ["user", "count"],
-    queryFn: () => getUserCount(),
-  });
-};
+import { PageResponse, Sort } from "@shared/model";
+import { QueryClient, useMutation, useQuery } from "@tanstack/react-query";
+import { getRoles, getUsersPaged, updateUserActive } from "../api/UserApi";
+import { UserFilter, UserListDto } from "@/features";
 
 export const createUseUsersQueryKey = (
-  isServerSide: boolean,
   page?: number,
   sort?: Sort,
   filter?: UserFilter
 ) => {
-  return isServerSide
-    ? [
-        "user",
-        page!,
-        sort ? `${sort.sortField}-${sort.sortOrder}` : "not-sorting",
-        filter ? `${filter.staffNumber}` : "not-filtering",
-      ]
-    : ["user"];
+  return [
+    "user",
+    page!,
+    sort ? `${sort.sortField}-${sort.sortOrder}` : "not-sorting",
+    filter ? `f=${filter.staffNumber}` : "not-filtering",
+  ];
 };
 
-export const useUsers = (
-  isServerSide: boolean,
+export const useUsers = (page?: number, sort?: Sort, filter?: UserFilter) => {
+  return useQuery({
+    queryKey: createUseUsersQueryKey(page!, sort, filter),
+    queryFn: () => getUsersPaged(page!, sort, filter),
+  });
+};
+
+export const useUpdateActive = (
+  queryClient: QueryClient,
   page?: number,
   sort?: Sort,
-  filter?: UserFilter,
-  size?: number
+  filter?: UserFilter
 ) => {
-  if (isServerSide) {
-    return useSuspenseQuery({
-      queryKey: createUseUsersQueryKey(isServerSide, page!, sort, filter),
-      queryFn: () => getUsersPaged(page!, sort, filter),
-    });
-  } else {
-    return useSuspenseQuery({
-      queryKey: ["user"],
-      queryFn: () => getUsers(size!),
-    });
-  }
-};
-
-export const useUpdateActive = (onMutate: (userId: number) => void) => {
   return useMutation({
     mutationFn: (userId: number) => updateUserActive(userId),
-    onMutate,
+    onMutate: (userId: number) => {
+      queryClient.setQueryData<PageResponse<UserListDto>>(
+        createUseUsersQueryKey(page, sort, filter),
+        (oldData) => {
+          if (!oldData) return oldData;
+
+          return {
+            ...oldData,
+            data: oldData.content.map((user) =>
+              user.userId === userId
+                ? { ...user, isActive: !user.isActive }
+                : user
+            ),
+          };
+        }
+      );
+    },
   });
 };
