+++ packages/wms/src/features/user/ui/user-table/index.ts
@@ -1,2 +1,1 @@
 export * from "./UserTable";
-export * from "./UserTableWrapper";
